% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_analysis.R
\name{dimReduct}
\alias{dimReduct}
\title{dimReduct: Dimensional Reduction}
\usage{
dimReduct(
  object,
  color.variable1,
  assay.data = "q_norm",
  do.log = TRUE,
  color.variable2 = NULL,
  shape.variable = NULL,
  point.size = 1,
  point.alpha = 1,
  symbol.size = 2,
  text.size = 16,
  print.plots = FALSE
)
}
\arguments{
\item{object}{of class \emph{NanoStringGeoMxSet}}

\item{color.variable1}{categorical variable to be used for coloring points}

\item{assay.data}{\emph{assayDataElement} of the \emph{NanoStringGeoMxSet}-class object to
be used for dimensional reductions}

\item{do.log}{boolean to apply log2 transformation to the data to be used
in the dimensional reduction algorithms (TRUE by default)}

\item{color.variable2}{categorical variable to be used for coloring points
(optional)}

\item{shape.variable}{categorical variable to be used for the shape of points
(optional)}

\item{point.size}{point size}

\item{point.alpha}{point color transparency}

\item{symbol.size}{symbols size in legends}

\item{text.size}{base font size}

\item{print.plots}{boolean to display plots or set to FALSE otherwise; \emph{ggplot}
objects are returned in either case}
}
\value{
A named list containing the \emph{NanoStringGeoMxSet}-class object with the
dimensional reduction coordinates added ("object") and a list of \emph{ggplot}
objects ("plot")
}
\description{
Reduces data to two dimensions using PCA, tSNE, and UMAP. Adds
the sample embeddings of each dimensional reduction technique to the input
object (\emph{phenoData}) and plots them on 2D scatter plots.
}
\details{
If both color.variable1 and color.variable2 are defined, points in
the scatter plots are colored by a factor which combines the levels of
these two variables.
}
\examples{
\dontrun{
# Default plots
output <- dimReduct(object = object, color.variable1="region", 
print.plots=TRUE)
# Plot tSNE
output <- dimReduct(object = object, color.variable1="region",
print.plots = FALSE)
figure <- output$plot$tSNE
print(figure)
# Color by two variables
output <- dimReduct(object = object, color.variable1="region",
color.variable2="class")
}
}
\seealso{
\code{\link{stats::prcomp}} \code{\link{Rtsne::Rtsne}}
\code{\link{umap::umap}}
}
