% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filtering}
\alias{filtering}
\title{Filtering NanoStringGeoMxSet dataset}
\usage{
filtering(object, pkc.file, loq.cutoff, loq.min, cut.segment, goi)
}
\arguments{
\item{object}{A NanoStringGeoMxSet dataset}
}
\value{
A list containing the ....
}
\description{
Function for filtering NanoStringGeoMxSet dataset
}
\details{
Normalize Nano String Digital Spatial Profile

This function will run various filtering parameters for NanoStringGeoMxSet datasets
}
