% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_analysis.R
\name{GeomxDimensionReduction}
\alias{GeomxDimensionReduction}
\title{Ingest Nano String Digital Spatial Profile study}
\usage{
GeomxDimensionReduction(
  Object,
  point.size = 1,
  point.alpha = 1,
  color.variable
)
}
\arguments{
\item{Object}{NanoStringGeoMxSet object with normalized data}

\item{point.size}{point size}

\item{point.alpha}{point color transparency}

\item{color.variable}{categorical variable to be used for coloring points}
}
\description{
Ingest Nano String Digital Spatial Profile study
}
