% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{geomxNorm}
\alias{geomxNorm}
\title{Normalization function}
\usage{
geomxNorm(object, norm)
}
\arguments{
\item{object}{A NanoStringGeoMxSet dataset}

\item{norm}{A vector with options of c(q3 or neg)}
}
\value{
A list containing the ggplot grid, a boxplot, an normalized dataframe.
}
\description{
Returns a dataset that are normalized with various normalized options
showing data after normalization.
}
\details{
normalization function that takes in gene expression data from
NanostringGeoMxSet, outputs a normalized NanostringGeoMxSet
}
