% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_preprocessing.R
\name{qcProc}
\alias{qcProc}
\title{qcProc: Quality Control and Preprocessing}
\usage{
qcProc(
  object,
  min.segment.reads = 1000,
  percent.trimmed = 80,
  percent.stitched = 80,
  percent.aligned = 80,
  percent.saturation = 50,
  min.nuclei = 200,
  min.area = 16000,
  min.negative.count = 10,
  max.ntc.count = 60,
  min.probe.ratio = 0.1,
  outlier.test.alpha = 0.01,
  percent.fail.grubbs = 20,
  remove.local.outliers = FALSE,
  shift.counts.zero = TRUE,
  print.plots = FALSE
)
}
\arguments{
\item{object}{object of class \emph{NanoStringGeoMxSet} with raw data and metadata
supplied}

\item{min.segment.reads}{minimum number of reads}

\item{percent.trimmed}{minimum \% of reads trimmed}

\item{percent.stitched}{minimum \% of reads stitched}

\item{percent.aligned}{minimum \% of reads aligned}

\item{percent.saturation}{minimum sequencing saturation}

\item{min.nuclei}{minimum # of nuclei estimated}

\item{min.area}{minimum segment area}

\item{min.negative.count}{minimum negative control counts}

\item{max.ntc.count}{maximum counts observed in NTC well}

\item{min.probe.ratio}{numeric between 0 and 1 to flag probes that have
(geomean probe in all segments) / (geomean probes within target) less than or
equal to this ratio}

\item{outlier.test.alpha}{significance cut-off for Grubb's test to flag
outlier probes}

\item{percent.fail.grubbs}{numeric to flag probes that fail Grubb's test in
greater than or equal this percent of segments}

\item{remove.local.outliers}{boolean to determine if local outliers should be
excluded from exprs matrix}

\item{shift.counts.zero}{boolean to shift 0 counts (if TRUE shift 0s by 1,
otherwise shift all counts by 1)}

\item{print.plots}{boolean to display plots or set to FALSE otherwise; ggplot
objects are returned in either case}
}
\value{
A list of two objects:
\itemize{
\item QC-filtered \emph{NanoStringGeoMxSet} object with gene-level targets as features
and the \emph{QCFlags} data.frame appended to \emph{protocolData} ("object")
\item QC plots in a list of ggplot objects ("plot")
\item Summary tables for segment, probe, and gene filtering ("table")
}
}
\description{
Selects ROI/AOI segments and target probes based on quality
control and then generates gene-level count data.
}
\details{
A probe is removed globally from the dataset if either it does not
reach the \emph{min.probe.ratio} threshold or the probe is an outlier according to
Grubb’s test in at least \emph{percent.fail.grubbs} of segments. A probe is
removed locally (from a given segment) if the probe is an outlier in that
segment and \emph{remove.local.outliers} is set to TRUE.

The count for any gene with multiple probes per segment is calculated as the
geometric mean of those probes.
}
