% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_deconvolution.R
\name{spatial_deconvolution}
\alias{spatial_deconvolution}
\title{Estimate cell composition across spatial samples using information from a reference matrix}
\usage{
spatial_deconvolution(dsp_qnorm, dsp_negnorm, ref_mtx, ref_annot)
}
\arguments{
\item{dsp_qnorm}{normalized expression matrix (Gene x DSP_Samples).}

\item{dsp_negnorm}{matrix of expected background for all data points in the normalized data matrix}

\item{ref_mtx}{expression matrix (Gene x Reference_Samples)}

\item{ref_annot}{annotated data.frame with CellID and LabeledCellType}
}
\value{
A list containing the results of spatial deconvolution, res$beta contains matrix of estimated cell abundances, heatmap of estimated cell abundances
}
\description{
Estimate cell composition across spatial samples using information from a reference matrix
}
