% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_deconvolution.R
\name{spatialDeconvolution}
\alias{spatialDeconvolution}
\title{Estimate cell composition across spatial samples using information from a custom cell signature matrix}
\usage{
spatialDeconvolution(
  dsp.obj,
  norm.expr.type = "",
  ref.mtx,
  ref.annot,
  cell.id = "CellID",
  celltype.col = "LabeledCellType",
  normalize = FALSE,
  min.cell.num = 0,
  min.genes = 10
)
}
\arguments{
\item{dsp.obj}{Nanostring object containing normalized gene expression data}

\item{norm.expr.type}{Name of the slot containing normalized gene expression data}

\item{ref.mtx}{expression matrix (Gene x Reference_Samples)}

\item{ref.annot}{annotated data.frame with cell.id and LabeledCellType, will be used along with ref.mtx to generate custom signature matrix}

\item{cell.id}{column of annotated data.frame containing cell barcode or name}

\item{celltype.col}{column of annotated data.frame containing cell identity information}

\item{normalize}{scale profile matrix gene expression according to gene count}

\item{min.cell.num}{minimum number of cells (within a type) required to generate signature matrix}

\item{min.genes}{filters cells according to minimum number of genes expressed}
}
\value{
A list res containing the results of spatial deconvolution,
res$beta: matrix of estimated cell abundances
res$X: cell signature (gene x celltype) matrix used to generate deconvolution results
res$yhat, res$resids: fitted values and log2-scale residuals from deconvolution, can be used to measure observations' goodness of fit
res$prop_of_all: res$beta rescaled to proportion of celltype across each sample
heatmap / barplot of estimated cell abundances in data
}
\description{
Estimate cell composition across spatial samples using information from a custom cell signature matrix
}
