% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_design.R
\name{studyDesign}
\alias{studyDesign}
\title{studyDesign: Study Design}
\usage{
studyDesign(
  dcc.files,
  pkc.files,
  pheno.data.file,
  pheno.data.sheet = "Template",
  pheno.data.dcc.col.name = "Sample_ID",
  protocol.data.col.names = c("aoi", "roi"),
  experiment.data.col.names = c("panel"),
  slide.name.col = "slide name",
  class.col = "class",
  region.col = "region",
  segment.col = "segment",
  area.col = "area",
  nuclei.col = "nuclei",
  sankey.exclude.slide = FALSE,
  segment.id.length = 4
)
}
\arguments{
\item{dcc.files}{A character vector containing the paths to the DCC files.}

\item{pkc.files}{A character vector representing the path to the
corresponding PKC file}

\item{pheno.data.file}{An optional character string representing the path to
the corresponding phenotypic excel data file. It is recommended to use the
Lab Worksheet in the exact order samples are provided in.}

\item{pheno.data.sheet}{An optional character string representing the excel
sheet name containing the phenotypic data.}

\item{pheno.data.dcc.col.name}{Character string identifying unique sample
identifier column in phenoDataFile.}

\item{protocol.data.col.names}{Character list of column names from
phenoDataFile containing data about the experimental protocol or
sequencing data.}

\item{experiment.data.col.names}{Character list of column names from
phenoDataFile containing data about the experiment's meta-data.}

\item{slide.name.col}{The name of the field that contains the slide names}

\item{class.col}{The name of the field that contains the class annotation}

\item{region.col}{The name of the field that contains the region annotation}

\item{segment.col}{The name of the field that contains the segment name}

\item{area.col}{The name of the field that contains the segment area}

\item{nuclei.col}{The name of the field that contains the nuclei number}

\item{sankey.exclude.slide}{A toggle for including the slide name in the
Sankey plot}

\item{segment.id.length}{The number of characters to use from each of the
annotation fields class, region, and segment to create the segment ID}
}
\value{
A list containing the NanoString Object and the Sankey plot.
}
\description{
Combine Nanostring Digital Spatial Profile read count and
annotation files into a GeoMX object
}
\details{
\code{studyDesign} returns a Sankey Plot and a GeoMX object.
For the function to run properly the annotation excel file must have the
specific field names: slide name', class, segment, region
and the corresponding fields that can be renamed: Sample_ID, aoi, roi, panel
}
