% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_plot.R
\name{violin_function}
\alias{violin_function}
\title{View gene expression expressed across groups and optionally across (broader) class}
\usage{
violin_function(
  data = target_demoData,
  genes = c(),
  group = "CellType",
  facet_by = NULL,
  expr_type = "q_norm"
)
}
\arguments{
\item{genes}{which genes should be plotted on the y-axis of the violin plot}

\item{group}{which groups should be plotted on the x axis of the violin plot}

\item{facet_by}{broader category of group, if specified, the violin plots will be split further based on this category}

\item{expr_type}{name of normalized expression data slot}
}
\value{
Violin plot(s) of genes to view across groups and (optionally) faceted across classes
}
\description{
View gene expression expressed across groups and optionally across (broader) class
}
