% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_plot.R
\name{violinPlot}
\alias{violinPlot}
\title{View gene expression expressed across groups and optionally across (broader) class}
\usage{
violinPlot(
  data = target_demoData,
  expr.type = "q_norm",
  genes = c(),
  group = "CellType",
  facet.by = NULL
)
}
\arguments{
\item{expr.type}{name of normalized expression data slot}

\item{genes}{which genes should be plotted on the y-axis of the violin plot}

\item{group}{which groups should be plotted on the x axis of the violin plot}

\item{facet.by}{broader category of group, if specified, the violin plots will be split further based on this category}
}
\value{
Violin plot(s) of genes to view across groups and (optionally) faceted across classes
}
\description{
View gene expression expressed across groups and optionally across (broader) class
}
