% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression_analysis.R
\name{DiffExpr}
\alias{DiffExpr}
\title{Run a linear mixed model on GeoMxSet}
\usage{
DiffExpr(
  object,
  analysisType,
  regions,
  groups,
  slideCol = "slide name",
  groupCol = "class",
  regionCol = "region",
  element = "log_q",
  multiCore = TRUE,
  nCores = 1,
  pAdjust = BY,
  pairwise = NULL,
  fclim = 1.2,
  pvallim1 = 0.05,
  pvallim2 = 0.01
)
}
\arguments{
\item{object}{name of the NanoStringGeoMxSet to perform DE analysis on}

\item{analysisType}{Analysis type either "Between Groups" or "Within Groups"}

\item{regions}{One or more regions of interest (needs to be variables within regionCol)}

\item{groups}{One or more groups of interest (needs to be variables within groupCol)}

\item{slideCol}{column for slide name}

\item{groupCol}{column for group (groups are usually found in different slides)}

\item{regionCol}{column for region (regions are usually found within a slide)}

\item{element}{assayDataElement of the geoMxSet object to run the DE on}

\item{multiCore}{= TRUE, set to TRUE to use multiCore, FALSE to run in cluster mode}

\item{nCores}{= 1, number of cores to use, set to 1 if running in serial mode}

\item{pAdjust}{= "BY" method for p-value adjustment}

\item{pairwise}{boolean to calculate least-square means pairwise differences}

\item{fclim}{= 1.2, Fold Change limit for summarizing genes of interest}
}
\value{
a list containing mixed model output data frame, grid tables for samples and summary of genelists
}
\description{
Run a linear mixed model on GeoMxSet
}
