% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression_analysis.R
\name{diffExpr}
\alias{diffExpr}
\title{Run a linear mixed model on GeoMxSet}
\usage{
diffExpr(
  object,
  analysis.type,
  groups,
  group.col,
  regions,
  region.col,
  slide.col = "slide_name",
  element = "q_norm",
  multi.core = TRUE,
  n.cores = 1,
  p.adjust = "fdr",
  pairwise = TRUE,
  fc.lim = 1.2,
  pval.lim.1 = 0.05,
  pval.lim.2 = 0.01
)
}
\arguments{
\item{object}{Name of the NanoStringGeoMxSet to perform DE analysis on}

\item{analysis.type}{Analysis type either "Between Groups" or "Within Groups"}

\item{groups}{One or more groups of interest (needs to be variables within
group.col)}

\item{group.col}{Column for group (groups are usually found in different
slides)}

\item{regions}{One or more regions of interest (needs to be variables within
region.col)}

\item{region.col}{Column for region (regions are usually found within a
slide)}

\item{slide.col}{Column for slide name (default is "slide name")}

\item{element}{assayDataElement of the geoMxSet object to run the DE on}

\item{multi.core}{Set to TRUE to use multicore, FALSE to run in cluster mode
(default is TRUE)}

\item{n.cores}{Number of cores to use, set to 1 if running in serial mode
(default is 1)}

\item{p.adjust}{Method to use for pvalue adjustment. Choices are "holm",
"hochberg","hommel","bonferroni","BH","BY","fdr","none". (default is "BH")}

\item{pairwise}{Boolean to calculate least-square means pairwise differences
(default is TRUE)}

\item{fc.lim}{Fold Change limit for summarizing genes of interest (default
is 1.2)}

\item{pval.lim.1}{P-value limit for summarizing differentially expressed
(DEG) genes, usually more stringent (default is 0.05)}

\item{pval.lim.2}{P-value limit for summarizing differentially expressed
(DEG) genes, usually more lenient (default is 0.10)}
}
\value{
a list containing mixed model output data frame, grid tables for
samples used in analysis and summary of significant genelists
}
\description{
diffExpr returns a DEG table with fold changes and p-values and
an image containing 2 tables, one showing the breakdown of number of samples
used in the analysis and a summary table for number of differentially
expressed genes (DEG) given manually set fold-change and p-value thresholds.
}
\details{
This function will run mixedModelDE from the GeoMxTools package.
}
