% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression_analysis.R
\name{DiffExpr}
\alias{DiffExpr}
\title{Run a linear mixed model on GeoMxSet}
\usage{
DiffExpr(
  object,
  element,
  analysisType,
  regions,
  groups,
  slideCol,
  classCol,
  fclim,
  multiCore,
  nCores,
  pAdjust,
  pairwise
)
}
\arguments{
\item{object}{name of the NanoStringGeoMxSet to perform DE analysis on}

\item{element}{assayDataElement of the geoMxSet object to run the DE on}

\item{analysisType}{Analysis type either "Between Groups" or "Within Groups"}

\item{regions}{vector of regions of interest}

\item{groups}{vector of groups to compare}

\item{slideCol}{column for slide name}

\item{classCol}{column for class}

\item{fclim}{= 1.2, Fold Change limit for summarizing genes of interest}

\item{multiCore}{= TRUE, set to TRUE to use multiCore, FALSE to run in cluster mode}

\item{nCores}{= 1, number of cores to use, set to 1 if running in serial mode}

\item{pAdjust}{= "BY" method for p-value adjustment}

\item{pairwise}{boolean to calculate least-square means pairwise differences}
}
\value{
a list containing mixed model output data frame, grid tables for samples and summary of genelists
}
\description{
Run a linear mixed model on GeoMxSet
}
