% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised_analysis.R
\name{DimReduct}
\alias{DimReduct}
\title{DimReduct: Dimensional Reduction}
\usage{
DimReduct(
  object,
  color.variable1,
  color.variable2 = NULL,
  shape.variable = NULL,
  point.size = 1,
  point.alpha = 1
)
}
\arguments{
\item{object}{of class NanoStringGeoMxSet with q_norm slot present}

\item{color.variable1}{categorical variable to be used for coloring points (required)}

\item{color.variable2}{categorical variable to be used for coloring points (optional)}

\item{shape.variable}{categorical variable to be used for the shape of points (optional)}

\item{point.size}{point size}

\item{point.alpha}{point color transparency}
}
\value{
A named list containing the NanoStringGeoMxSet-class object with the dimensional reduction coordinates ("dsp.object") and ggplot2 plots ("plot.list")
}
\description{
Reduces data to 2-dimensions using PCA, tSNE, and UMAP. Adds the sample embeddings of each dimensional reduction technique to the input object (phenoData) and plots them on 2D scatter plots.
}
\details{
If both color.variable1 and color.variable2 are defined, points in the scatter plots are colored by a factor which combines the levels of these two variables.
}
\seealso{
\link[stats]{prcomp} \link[Rtsne]{Rtsne}  \link[umap]{umap}
}
