% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filtering}
\alias{filtering}
\title{Normalize Nano String Digital Spatial Profile}
\usage{
filtering(Data, dsp_obj, PKCS, LOQcutoff, LOQmin, CutSegment, GOI)
}
\arguments{
\item{Data}{A NanoStringGeoMxSet dataset}
}
\value{
A list containing the ....
}
\description{
Normalize Nano String Digital Spatial Profile
}
