% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{GeoMxNorm}
\alias{GeoMxNorm}
\title{Normalize Nano String Digital Spatial Profile}
\usage{
GeoMxNorm(Data, Norm)
}
\arguments{
\item{Data}{A NanoStringGeoMxSet dataset}

\item{Norm}{A vector with options of c(quant or neg)}
}
\value{
A list containing the ggplot grid, a boxplot, an normalized dataframe.
}
\description{
Normalize Nano String Digital Spatial Profile
}
