% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{HeatMap}
\alias{HeatMap}
\title{NanoString Digital Spatial Profile study}
\usage{
HeatMap(
  target.data,
  ngenes,
  image.width,
  image.height,
  image.resolution,
  image.filename,
  heatmap.color
)
}
\arguments{
\item{target.data}{Normalized targetData S4 object input file.}

\item{ngenes}{Number of high CV genes to cluster and plot}

\item{image.width}{image width}

\item{image.height}{image height}

\item{image.resolution}{image resolution}

\item{image.filename}{image filename}

\item{heatmap.color}{colors of heatmap

#' @importFrom NanoStringNCTools assayDataApply}
}
\value{
A list containing the plot genes data matrix, and the heatmap plot.
}
\description{
\enumerate{
\item Exploring the normalized data by calculating the coefficient of variation (CV) for each gene (g) using the formula CVg=SDg/meang.
\item We then identify genes with high CVs that should have large differences across the various profiled segments.
\item This unbiased approach can reveal highly variable genes across the study.
\item We plot the results using unsupervised hierarchical clustering, displayed as a heatmap.
}
}
