% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_preprocessing.R
\name{QcProc}
\alias{QcProc}
\title{QcProc: Quality Control and Preprocessing}
\usage{
QcProc(
  object,
  minSegmentReads = 1000,
  percentTrimmed = 80,
  percentStitched = 80,
  percentAligned = 80,
  percentSaturation = 50,
  minNegativeCount = 10,
  maxNTCCount = 1000,
  minNuclei = 100,
  minArea = 5000,
  minProbeRatio = 0.1,
  outlierTestAlpha = 0.01,
  percentFailGrubbs = 20,
  removeLocalOutliers = TRUE
)
}
\arguments{
\item{object}{object of class \emph{NanoStringGeoMxSet} with raw data and metadata supplied}

\item{minSegmentReads}{minimum number of reads}

\item{percentTrimmed}{minimum \% of reads trimmed}

\item{percentStitched}{minimum \% of reads stitched}

\item{percentAligned}{minimum \% of reads aligned}

\item{percentSaturation}{minimum sequencing saturation}

\item{minNegativeCount}{minimum negative control counts}

\item{maxNTCCount}{maximum counts observed in NTC well}

\item{minNuclei}{minimum # of nuclei estimated}

\item{minArea}{minimum segment area}

\item{minProbeRatio}{numeric between 0 and 1 to flag probes that have (geomean probe in all segments) / (geomean probes within target) less than or equal to this ratio}

\item{outlierTestAlpha}{significance cut-off for Grubb's test to flag outlier probes}

\item{percentFailGrubbs}{numeric to flag probes that fail Grubb's test in greater than or equal this percent of segments}

\item{removeLocalOutliers}{boolean to determine if local outliers should be excluded from exprs matrix}
}
\value{
A list of two objects: QC-filtered \emph{NanoStringGeoMxSet} object with gene-level targets as features and the \emph{QCFlags} data frame appended to \emph{protocolData} and QC plots
}
\description{
Selects ROI/AOI segments and target probes based on quality control and then generates gene-level count data.
}
\details{
A probe is removed globally from the dataset if either it does not reach the \emph{minProbeRatio} threshold or the probe is an outlier according to Grubb’s test in at least \emph{percentFailGrubbs} of segments. A probe is removed locally (from a given segment) if the probe is an outlier in that segment and \emph{removeLocalOtliers} is set to TRUE (default).

The count for any gene with multiple probes per segment is calculated as the geometric mean of those probes.
}
