% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_deconvolution.R
\name{spatial_deconvolution}
\alias{spatial_deconvolution}
\title{Estimate cell composition across spatial samples using information from a custom cell signature matrix}
\usage{
spatial_deconvolution(
  dsp_qnorm,
  dsp_negnorm,
  ref_mtx,
  ref_annot,
  CellID = "CellID",
  cellTypeCol = "LabeledCellType",
  normalize = FALSE,
  minCellNum = 0,
  minGenes = 10
)
}
\arguments{
\item{dsp_qnorm}{normalized expression matrix (Gene x DSP_Samples).}

\item{dsp_negnorm}{matrix of expected background for all data points in the normalized data matrix}

\item{ref_mtx}{expression matrix (Gene x Reference_Samples)}

\item{ref_annot}{annotated data.frame with CellID and LabeledCellType, will be used along with ref_mtx to generate custom signature matrix}

\item{CellID}{column of annotated data.frame containing cell barcode or name}

\item{normalize}{scale profile matrix gene expression according to gene count}

\item{minCellNum}{minimum number of cells (within a type) required to generate signature matrix}

\item{minGenes}{filters cells according to minimum number of genes expressed}

\item{celltypeCol}{column of annotated data.frame containing cell identity information}
}
\value{
A list res containing the results of spatial deconvolution,
res$beta: matrix of estimated cell abundances
res$X: cell signature (gene x celltype) matrix used to generate deconvolution results
res$yhat, res$resids: fitted values and log2-scale residuals from deconvolution, can be used to measure observations' goodness of fit
res$prop_of_all: res$beta rescaled to proportion of celltype across each sample
heatmap / barplot of estimated cell abundances in data
}
\description{
Estimate cell composition across spatial samples using information from a custom cell signature matrix
}
