% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_deconvolution.R
\name{spatialDeconvolution}
\alias{spatialDeconvolution}
\title{Spatial Deconvolution

Helper functions comes from
https://bioconductor.org/packages/release/bioc/vignettes/SpatialDecon/inst/doc/SpatialDecon_vignette_NSCLC.html}
\usage{
spatialDeconvolution(
  object,
  expr.type,
  ref.mtx,
  ref.annot,
  cell.id.col,
  celltype.col,
  group.by = NULL,
  out.directory = NULL,
  matrix.name = "customDSPmtx",
  normalize = FALSE,
  min.cell.num = 0,
  min.genes = 10,
  discard.celltype = FALSE
)
}
\arguments{
\item{object}{Nanostring object containing normalized gene expression data}

\item{expr.type}{Name of slot containing normalized gene expression data}

\item{ref.mtx}{Reference expression matrix (Gene x Reference_Samples)}

\item{ref.annot}{Reference data.frame with cell.id and celltype information}

\item{cell.id.col}{Column of data.frame containing cell.id.col info}

\item{celltype.col}{Column of data.frame containing celltype info}

\item{out.directory}{Path to desired output directory, set to NULL if matrix
should not be written}

\item{matrix.name}{Name given to deconvolution signature matrix}

\item{normalize}{Scale profile matrix gene expression according to gene count}

\item{min.cell.num}{Prevent deconvolution of celltype(s) if number of
corresponding cells is below this threshold}

\item{min.genes}{Filter cells based on minimum number of genes expressed}

\item{discard.celltype}{Remove any celltype(s) that is not of interest}
}
\value{
A list dsp.data containing the results of spatial deconvolution,
res$beta: matrix of estimated cell abundances
res$X: cell signature (gene x celltype) matrix used to generate
deconvolution results
res$yhat, res$resids: fitted values and log2-scale residuals from
deconvolution, can be used to measure goodness of fit
res$prop_of_all: res$beta scaled to proportion of celltype across sample
figures: heatmap and barplot of cell profile matrix and estimated
cell abundances in data
}
\description{
spatialDeconvolution estimate cell composition across DSP
samples from reference expression matrix
}
\details{
Uses Nanostring developed functions to compute estimated cell
fractions in DSP samples. Allows for users to group samples based
on metadata information
}
