% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_deconvolution.R
\name{spatialDeconvolution}
\alias{spatialDeconvolution}
\title{Spatial Deconvolution}
\usage{
spatialDeconvolution(
  object,
  expr.type = "q_norm",
  prof.mtx,
  clust.rows = TRUE,
  clust.cols = TRUE,
  group.by = "none",
  plot.fontsize = 5,
  use.custom.prof.mtx = FALSE,
  discard.celltype = FALSE,
  normalize = FALSE,
  min.cell.num = 0,
  min.genes = 10,
  ref.mtx,
  ref.annot,
  cell.id.col = "CellID",
  celltype.col = "LabeledCellType"
)
}
\arguments{
\item{object}{Nanostring object containing normalized gene expression data}

\item{expr.type}{Name of slot containing normalized gene expression data}

\item{prof.mtx}{Use stored profile matrix}

\item{clust.rows}{Cluster rows in heatmap (Default: TRUE)}

\item{clust.cols}{Cluster columns in heatmap (Default: TRUE)}

\item{group.by}{Organize heatmap / barplot columns by metadata group
(Default: "none")}

\item{plot.fontsize}{Set size of labels on all figures (Default: 5)}

\item{discard.celltype}{Remove any celltype(s) that is not of interest
(Default: FALSE)}

\item{normalize}{Scale profile matrix gene expression according to gene count
(Default: FALSE)}

\item{min.cell.num}{Prevent deconvolution of celltype(s) if number of
corresponding cells is below this threshold (Default: 0)}

\item{min.genes}{Filter cells based on minimum number of genes expressed
(Default: 10)}

\item{ref.mtx}{Custom reference expression matrix (Gene x Reference_Samples)}

\item{ref.annot}{Custom reference data.frame with cell.id and celltype
information}

\item{cell.id.col}{Column of data.frame containing cell.id.col info}

\item{celltype.col}{Column of data.frame containing celltype info}

\item{use.custom.matrix}{Generate custom profile matrix (Default: FALSE)}
}
\value{
A list dsp.data containing the results of spatial deconvolution,
res$beta: matrix of estimated cell abundances
res$X: cell signature (gene x celltype) matrix used to generate
deconvolution results
res$yhat, res$resids: fitted values and log2-scale residuals from
deconvolution, can be used to measure goodness of fit
res$prop_of_all: res$beta scaled to proportion of celltype across sample
figures: heatmap and barplot of cell profile matrix and estimated
cell abundances in data
}
\description{
spatialDeconvolution estimate cell composition across DSP
samples from reference expression matrix
}
\details{
Helper functions comes from
https://bioconductor.org/packages/release/bioc/vignettes/SpatialDecon/inst/doc/SpatialDecon_vignette_NSCLC.html

Uses Nanostring developed functions to compute estimated cell
fractions in DSP samples. Allows for users to group samples based
on metadata information
}
