% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_design.R
\name{StudyDesign}
\alias{StudyDesign}
\title{Combine Nanostring Digital Spatial Profile read count and annotation files into a GeoMX object}
\usage{
StudyDesign(
  dcc.files,
  pkc.files,
  pheno.data.file,
  pheno.data.sheet = "Template",
  pheno.data.dcc.col.name = "Sample_ID",
  protocol.data.col.names = c("aoi", "roi"),
  experiment.data.col.names = c("panel")
)
}
\arguments{
\item{dccFiles}{A character vector containing the paths to the DCC files.}

\item{pkcFiles}{A character vector representing the path to the corresponding PKC file}

\item{phenoDataFile}{An optional character string representing the path to the corresponding phenotypic excel data file. It is recommended to use the Lab Worksheet in the exact order samples are provided in.}

\item{phenoDataSheet}{An optional character string representing the excel sheet name containing the phenotypic data.}

\item{phenoDataDccColName}{Character string identifying unique sample identifier column in phenoDataFile.}

\item{protocolDataColNames}{Character list of column names from phenoDataFile containing data about the experimental protocol or sequencing data.}

\item{experimentDataColNames}{Character list of column names from phenoDataFile containing data about the experiment's meta-data.}
}
\value{
A list containing the NanoString Object and the Sankey plot.
}
\description{
\code{StudyDesign} returns a Sankey Plot and a GeoMX object
}
\details{
For the function to run properly the annotation excel file must have the specific field names:
'slide name', class, segment, region

and the corresponding fields that can be renamed:
Sample_ID, aoi, roi, panel
}
