% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_design.R
\name{studyDesign}
\alias{studyDesign}
\title{studyDesign: Study Design}
\usage{
studyDesign(
  dcc.files,
  pkc.files,
  pheno.data.file,
  pheno.data.sheet = "Template",
  pheno.data.dcc.col.name = "Sample_ID",
  protocol.data.col.names = c("aoi", "roi"),
  experiment.data.col.names = c("panel")
)
}
\arguments{
\item{dcc.files}{A character vector containing the paths to the DCC files.}

\item{pkc.files}{A character vector representing the path to the
corresponding PKC file}

\item{pheno.data.file}{An optional character string representing the path to
the corresponding phenotypic excel data file. It is recommended to use the
Lab Worksheet in the exact order samples are provided in.}

\item{pheno.data.sheet}{An optional character string representing the excel
sheet name containing the phenotypic data.}

\item{pheno.data.dcc.col.name}{Character string identifying unique sample
identifier column in phenoDataFile.}

\item{protocol.data.col.names}{Character list of column names from
phenoDataFile containing data about the experimental protocol or
sequencing data.}

\item{experiment.data.col.names}{Character list of column names from
phenoDataFile containing data about the experiment's meta-data.}
}
\value{
A list containing the NanoString Object and the Sankey plot.
}
\description{
Combine Nanostring Digital Spatial Profile read count and
annotation files into a GeoMX object
}
\details{
\code{studyDesign} returns a Sankey Plot and a GeoMX object.
For the function to run properly the annotation excel file must have the
specific field names: slide name', class, segment, region
and the corresponding fields that can be renamed: Sample_ID, aoi, roi, panel
}
