% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_design.R
\name{StudyDesign}
\alias{StudyDesign}
\title{Ingest Nano String Digital Spatial Profile study}
\usage{
StudyDesign(
  dccFiles,
  pkcFiles,
  phenoDataFile,
  phenoDataSheet = "Template",
  phenoDataDccColName = "Sample_ID",
  protocolDataColNames = c("aoi", "roi"),
  experimentDataColNames = c("panel")
)
}
\arguments{
\item{dccFiles}{A character vector containing the paths to the DCC files.}

\item{pkcFiles}{A character vector representing the path to the corresponding PKC file}

\item{phenoDataFile}{An optional character string representing the path to the corresponding phenotypic excel data file. It is recommended to use the Lab Worksheet in the exact order samples are provided in.}

\item{phenoDataSheet}{An optional character string representing the excel sheet name containing the phenotypic data.}

\item{phenoDataDccColName}{Character string identifying unique sample identifier column in phenoDataFile.}

\item{protocolDataColNames}{Character list of column names from phenoDataFile containing data about the experimental protocol or sequencing data.}

\item{experimentDataColNames}{Character list of column names from phenoDataFile containing data about the experiment's meta-data.}
}
\value{
A list containing the NanoString Object and the Sankey plot.
}
\description{
Ingest Nano String Digital Spatial Profile study
}
