% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_plot.R
\name{violinPlot}
\alias{violinPlot}
\title{Violin plot of spatially profiled genes}
\usage{
violinPlot(object, expr.type, genes, group, facet.by = NULL)
}
\arguments{
\item{object}{Name of NanostringGeoMxSet to perform analysis on}

\item{expr.type}{Name of normalized expression data slot}

\item{genes}{Transcript(s) to be plotted on the y-axis of the violin plot}

\item{group}{Metadata group to be plotted on the x axis of the violin plot}

\item{facet.by}{Further subdivide group by this category (Default: NULL)}
}
\value{
an arranged grob of violin plots
}
\description{
Returns violin plot of user-input gene(s) expression across
groups, can optionally subdivide groups into a more specific
category
}
\details{
ggplot function that takes in gene expression data from
NanostringGeoMxSet, outputs the distribution of gene expression
values across groups. Requires normalized expression data and valid
group category to be present in object
}
